support `_initialize` export as well as `__wasm_call_ctors`
Per https://github.com/WebAssembly/tool-conventions/pull/203 and recent LLVM and
wasi-libc changes, libraries may export an `_initialize` function instead of
`__wasm_call_ctors`, and we should call whichever one we find.  Note that we
expect to find at most one of the two, and will return an error if both are
found.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>