refactor: expose fn for embedding component metadata
The logic for embedding component metadata as a custom section was
used only in the course of `wasm-tools component embed`, which made it
difficult to use from outside wasm-tools.

This commit refeactors the code that actually writes out the metadata
to a function that can be re-used and is exposed by wit-component.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>