Use `serde_derive` instead of the `derive` feature
I noticed recently that `wit-component` was relatively slow to compile
and it's going to be an introductory requirement for any Rust crates so
I wanted to optimize its build a little bit. This is the first
optimization which is to disable the `derive` feature of the `serde`
crate to enable build build pipelining opportunities.

This is the same as bytecodealliance/wasmtime#6917