wasm-encoder: Add `From<wasmparser::Foo> for Foo` impls
This adds a new optional dependency on `wasmparser` from `wasm-encoder`. When
the optional dependency is enabled, we add a bunch of `From` impls to convert
from the `wasmparser` type to the `wasm_encoder` type. This isn't 100%
exhaustive yet, but we can always add more impls as we need them. This was
enough for the conversions that Winliner wanted, fwiw.

Fixes #1115