wasmparser: Define core Wasm types in a macro
This sets the stage for a second set of these definitions that don't use `u32`
indices, but instead use something similar to `CoreTypeId`. This new set of
types will be used in validation and will support canonicalization.

At the same time, I've taken the liberty to rename `StructuralType` to
`CompositeType` to match the latest spec and to change `HeapType::Indexed` to
`HeapType::Concrete` for the same reason.