implement resource support for `wasm-compose`
The basic approach is as follows:

- Start with an empty resource remapping (i.e. no given pair of distinct
  `ResourceId`s are considered equivalent)

- For instance import and export pair we're considering connecting to each
  other, match up any resource types by name and optimistically produce a
  remapping that sets identically-named pairs equal to each other, provided that
  remapping does not contradict any previous remappings.

- If the above fails (either because a consistent remapping could not be created
  or because the instances were incompatible for other reasons), throw away the
  changes and reject the connection (which may mean giving up on the whole
  composition if we can't do what the user requested).  Otherwise, store the
  remapping changes and accept the connection.

- Once all such connections have been made, update the remapping to set all
  remaining, unconnected, identically-named resource imports in
  identically-named instances equal to each other.

- Finally, encode the output component using the remapping produced above.

Note that I'm using `RefCell` for interior mutability since the places I needed
to update the remapping only had a shared reference to the `CompsitionGraph`,
and refactoring them to provide unique references felt like overkill (and I've
heard a major refactor of `wasm-compose` is planned anyway, so perhaps this can
be revisited then).

I've brought in `im-rc` as a dependency since it makes rolling back tentative
changes trivial.  There may be more efficient approaches if that's a concern.

Per Alex's advice, I've changed a few types and traits in `wasmparser::types`
from `pub(crate)` to `pub` in order to use them in the algorithm described
above.