Update parsing of component model names
This commit removes the `wasmparser::ComponentExternName` type and
updates wasmparser to handle WebAssembly/component-model#263. Names are
now always stored as a strings and are not discriminated in the binary
format. This additionally refactors the internals of `wasmparser` to
handle parsed names differently and updates tools like `wit-component`
to handle the new API changes.

Test aren't passing yet but things are compiling at least.