Turn on semicolons-in-WIT by default (#1269)
This commit updates tooling to require semicolons and print semicolons
by default in WIT files. This is the next step of this rollout plan now
that a month or so has passed since the original change. Tooling still
supports optionally not parsing or printing semicolons, as configured
with `WIT_REQUIRE_SEMICOLONS=0`. If this continues to be smooth the
next step will be to remove support for no-semicolons in the future.