Fix some issues with future encodings (#1270)
Currently there's a few encoding changes in-flight in wasm-tools. Two
relevant ones here are:

* How WIT packages are encoded (`WIT_COMPONENT_ENCODING_V2`, #1252)
* How type information is embedded in core wasm files
  (`WIT_COMPONENT_NEW_ENCODE`, #1260)

Currently these don't end up playing well together. If the new WIT
package encoding is used then it breaks when combined with the new
metadata encoding. This commit seeks to rectify this situation.
Currently this wasn't previously tested on CI due to this particular
combination of flags not being on in the tests.

The fixes here largely amount to refactoring to delete some duplicated
code. Previously the v1/v2 split introduced an `encode_world` that was
defined in both modules, but the output was different for both modules
and only one would work when fed to decoding. This commit removes this
split and leaves only one canonical `encode_world` function since it
should be the same for both as well.

This then additionally updates `WIT_COMPONENT_NEW_ENCODE=1` to preserve
a lookalike format of a WIT package. Furthermore `decode_world` is
updated to expect this structure as well. The intention is that the
encoded metadata is as-if there was a single world in a WIT package.

These refactorings fix the prior buggy behavior between the two new
encodings. Note that this was only exposed through
`WIT_COMPONENT_NEW_ENCODE=1` which was only very recently added.
Previous builds using only `WIT_COMPONENT_ENCODING_V2=1` should continue
to work fine and produce the same output as before.