Enable new WIT encoding by default
This commit enables the new encoding for WIT packages in WebAssembly
components described in WebAssembly/component-model#248 and originally
implemented in #1252. Support for the new encoding has been in a
wasm-tools for a bit and it's also released with Wasmtime 14. This
switch means that infrastructure will start being exposed to it by
default now. Support for the old encoding remains to assist with interop
as well. In a release or two support for creating the old encoding will
be removed.