Enable the `gc/global.wast` spec test
This involved queuing up table init expressions for later validation because:

* Tables (and their init expressions) come before globals are defined in the
  binary format,
* constant expressions can contain `global.get`s of defined globals now,
* and therefore a table init can validly reference a global before it is
  defined.

This is not ideal, but is fine, since we have some other end-of-module checking
already in `validate_end`. However, that alone isn't good enough. Because
constant expressions can contain `ref.func $f` instructions, which need to
insert `$f` into the "referenced function" set, we need mutable access to the
`Arc<Module>` when processing constant expressions. So we need to drain the
unchecked table init expressions *before* processing each code body, when we
still have unique access to the `Arc<Module>`. But not all Wasm modules have
code sections, so at the same time, we still need to do the checking inside
`validate_end`. And when there was a code section, the queue of table init
expressions will have already been drained by the time we are in `validate_end`,
so it is okay that we no longer have unique access to the `Arc<Module>`.