Enable new format for embedding component types by default (#1308)
This commit enables the support for the new format of type information
embedded in custom sections for `wit-component` first implemented
in #1260. The new format minimizes type information to only that which
was bound for the `world` selected as opposed to all packages and
interfaces regardless of whether they're referenced or not.

A bug with this format was fixed with #1270 and shipped in Wasmtime 15
so there's at least one stable release that supports the old and the new
without bugs. This commit turns on the new format by default while
leaving in an environment variable to turn it off. This isn't expected
to have any practical breakage at this time.