fall back to `cabi_realloc` if import/export variants not found
This is useful when doing shared-everything linking and we have a library module
that exports `cabi_realloc`, but not the adapter-specific `cabi_import_realloc`
and `cabi_export_realloc` functions.

For context: this came up while adding Preview 2 support to `wasi-libc` using
bindings generated by `wit-bindgen`, which only exports `cabi_realloc`.  That
works fine if you're statically linking `libc.a` and using `wit-component` the
traditional way, but if you're using `libc.so` and shared-everything linking,
each .so module is treated as an "adapter", which meant `wit-component` was
looking for `cabi_import_realloc` and `cabi_export_realloc`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>