Sync wasmparser/wasm-smith on bulk-memory (#1326)
* Sync wasmparser/wasm-smith on bulk-memory

Historically in #410 fuzzing for `wasm-tools smith` was updated to
unconditionally enable the bulk-memory proposal if the reference-types
proposal was enabled. This doesn't prevent configuration on the CLI,
however, as exposed in #1324. This change undoes #410 effectively and
then syncs the `wasm-smith` crate's understanding of bulk-memory and
reference-types with `wasmparser`'s understanding.

I'm not actually sure if `wasmparser`'s understanding of
bulk-memory-vs-reference-types is correct. It was always a bit murky to
me precisely which instructions where in which proposal, so I've opted
here to take whatever `wasmparser` does today as the source of truth.
Any edits necessary to move items between proposals can be done in
`wasmparser` and then reflected in `wasm-smith`.

Closes #1324

* Review comments