Add copy of subset of `ref.{test,cast}` tests
Not all are passing yet, but this is more related to our type canonicalization
and match checking than `ref.{test,cast}`. So in the meantime, check that the
subset of the spec tests that are passing continue to pass. Once we get all of
them passing we can enable the spec tests and delete this partial copy.