Print the name of label targets
This commit updates `wasmprinter` to print the target label name for
instructions such as `br` and friends. This is done by maintaining a
stack matching the stack of blocks that are entered, and the relative
depth is looked up in this stack to then consult the name label name map
contained in the label name subsection.

Closes #841