Add `wasmparser` support for `ref.{test,cast}` (#1332)
* Add `wasmparser` support for `ref.{test,cast}`

This also splits `wasm-encoder::Intruction::Ref{Test,Cast}` into nullable and
non-nullable forms, since that more directly aligns with the binary format where
there are actually different opcodes for nullable vs non-nullable variations (as
opposed to a bit on some immediate or something).

* Add copy of subset of `ref.{test,cast}` tests

Not all are passing yet, but this is more related to our type canonicalization
and match checking than `ref.{test,cast}`. So in the meantime, check that the
subset of the spec tests that are passing continue to pass. Once we get all of
them passing we can enable the spec tests and delete this partial copy.

* more concise

* fill out todo message