Fix `ref.{test,cast}` to check against top types
The spec says that `ref.{test,cast} rt` are valid when `rt'` is on the stack and
`rt <: rt'`. We were checking this before, but not allowing `rt'` to upcast to
any of its super types. The solution is to immediately upcast it to its top
type, and then check `rt <: top(rt')`.

This gets the `type-subtyping.wast` test fully passing.