wasm-smith: Generate GC types
Doesn't add support for generating GC instructions yet.

Also doesn't add support for non-nullable references yet, as there are various
places where we expect a type to be defaultable and/or have a value of the type
we can construct immediately and nullability is a good short hand for
that. Eventually we should properly support non-nullable references though.