Update the exception-handling validator implementation (#1333)
* Update the exception-handling validator implementation

This commit updates the implementation of the exception-handling
proposal to the latest version of the proposal. This was already
implemented in the text format but all other crates needed updating. The
changes were:

* A new `exn` heap type is added for `exnref`. Currently this isn't a
  subtype with anything else, and this probably isn't correct but it's
  at least conservative for now.

* The `try`, `delegate`, `catch`, `catch_all`, and `rethrow`
  instructions were all removed as they're no longer present.

* New `try_table` and `throw_ref` instructions were added.

* Support for the name section subsection for tags has been added along
  with printing tag names in the text format.

* All exception-handling spec tests are now re-enabled.

* Update error message

* Restore old text format

Continue to parse the old exception opcodes both in the text and binary
format. Additionally print them too. Don't validate them, however, and
additionally don't restore the sugared paren-style parsing of the old
text format. This should enable the ability to print out old binaries
and reassemble them but will likely render handwritten examples
un-workable due to the removal of the parentheses-based form.

* No tests need skipping at this time

* Fix a bit of mixing old and new syntax