Remove support for encoding old component metadata format
This commit is a follow-up to #1260, #1270, and #1308. This is the next
step in the transition to updating how type information is encoded in
object files in core WebAssembly binaries when transforming them into a
component. The old format has been enabled by default since the 1.0.54
release of `wasm-tools` in late November. and has been supported since
the 1.0.51 release in early November.

Support has been preserved for continuing to emit the old format in case
any tooling accidentally hadn't updated but now feels like the right
time to remove the ability to emit the old format. The old format can
still be parsed if historical object files are in use and the hope is
that it's not too bad to leave that piece sticking around for awhile.