`wasm-smith`: Generate GC types (#1360)
* wasm-smith: Generate GC types

Doesn't add support for generating GC instructions yet.

Also doesn't add support for non-nullable references yet, as there are various
places where we expect a type to be defaultable and/or have a value of the type
we can construct immediately and nullability is a good short hand for
that. Eventually we should properly support non-nullable references though.

* Enable Wasm GC during fuzzing

* wasmprinter: Fix printing of supertypes

Was just printing the supertype index in a comment if there wasn't a name
before. Now it prints the name when available or the index when it isn't
available.

* Fix config of fuzzing of Wasm GC

* Fix compilation in rebase

* Remove unnecessary || and add TODO about feature flags