`wast`: Fix pointer provenance bug reported by MIRI (#1386)
* wast: Fix pointer provenance bug reported by MIRI

We need to re-derive the str pointer after moving the original `Box<str>` it was
derived from.

* wast: Remove all unsafe, use bumpalo for arena

This removes the single unsafe block from the `wast` crate. The unsafe block was
extending the lifetime of some data by moving into an ad-hoc arena. We replace
that usage with `bumpalo`, which is well-tested, vetted, and etc...

* Default all crates in the workspace to `deny(unsafe_code)`

A couple crates still contain unsafe code: one block in `wasmparser` and all
throughout the C API. These have comments explaining their safety.

The one unsafe block that used to exist in `wasm-smith` has been removed. It was
an optimization to avoid double UTF-8 validation when constructing a `str` from
bytes. We simply validate twice now, since that cost should be negligible in the
context of the full Wasm test case generation.

* cargo fmt