Use different limits for items in components (#1390)
Right now the validation limit for the number of memories or tables in a
component is quite small, only 100. This is the core wasm limit which is
intended to limit the number of definitions in core wasm, but in a
component items taking up the index space are just references, not
definitions. That means they don't need to have similarly small limits
and can instead be much larger.

Closes #1389