Fix bugs involving the combination of branching and subtyping (#1403)
* wasmparser: Fix behavior of branching to labels and subtyping

Addresses issues discussed in https://github.com/WebAssembly/gc/issues/516

* wasm-smith: Properly erase subtype information after conditional branches

When, for example, `[a i32]` is on the stack and we have a `br_if @label` where
`@label`'s type is `[b]` and `a <: b`, then whole `br_if` is typed `[b i32] ->
[b]`. Because `a <: b` the `br_if` is valid. However, `a <: b` does *not* mean
that the instruction results in `[a]`. It "erases" the `a` and pushes a `b`.

This addresses https://github.com/WebAssembly/gc/issues/516 in `wasm-smith`.