Remove usage of `AtomicU64` in `wasmparser`
Instead substitute this with `AtomicUsize` which panics on overflow to
catch any possible issues on 32-bit platforms. This is motivated by
https://github.com/rust-lang/rust/pull/120588 which is pulling
`wasmparser` into the Rust compiler and `AtomicU64` is not available on
all the platforms the compiler itself is built for.

I plan on backporting this to the `0.118.x` release track as well which
is what's used by the `object` crate currently to avoid the need for a
whole bunch of cascading updates.