Fold `MaybeInvalidModule` into `Config` (#1411)
* Fold `MaybeInvalidModule` into `Config`

This commit removes the top-level `MaybeInvalidModule` type and instead
folds it into a configuration option inside of `Config`. This has the
benefit of being able to still configure all the other `Config` options
during generating possibly-invalid modules at the cost that we can no
longer statically rule out success of `ensure_termination`. This error
is now propagated upwards and documented in a few places.

Closes #1410

* Make anyhow non-optional for wasm-smith

* Update fuzzers