Add a `semver-check` subcommand (#1407)
* Add a `semver-check` subcommand

This issue is intended to address #1399. The assumption is that this
will be used by WASI repositories (eventually) to ensure that the
current `main` branch of the repository is semver-compatible with the
last release of the repository.

This commit adds a new subcommand which looks like:

    wasm-tools component semver-check ./wit --prev old-world --new new-world

This will check that `new-world` is a semver-compatible evolution of
`new-world`. Internally this works the same way as the `targets`
subcommand where a component is synthesized on-the-fly which is then fed
to `wasmparser` for validation to determine whether it's valid or not.
Specifically a component of shape `old-world` is fed into an import of
shape `new-world`, effectively answering the question of whether a
component taking the "old" world can be instantiated with the "new"
world.

Some basic CLI tests are added here as well showcasing various bits and
bobs. Some further tooling will be needed to actually integrate this
into WASI repositories but this should be at least one chunk of the
necessary work.

Closes #1399

* Fix wit-component build with few features