Add pointer and length types to `WasmType`.
In anticipation of memory64, provenance in Rust, and guest bindings with
instrumented pointers, add `Pointer` and `Length` types to `WasmType`,
and use it in the ABI for list, strings, and argument/return value
pointers. Consumers that don't have anything special to do with these can
handle them both the same as `i32`.

And, because the variant type Canonical ABI can unify pointers with
integer types like `i64`, add a `Pointer64` type as well, which
represents a conceptual union of a pointer and an `i64`. Consumers that
don't have anything special to do with this type can handle it the same
as an `i64`.