Refactor generation of tables/elements in wasm-smith
This commit refactors `wasm-smith` and its generation of both table
types and element segments. The goal is to help generate modules of
shapes that Wasmtime does not currently support but should. Notably
constant expressions are allowed to use `global.get`, even in element
segments, and Wasmtime does not currently support this.

Furthermore this commit additionally enables generating tables with
initialization expressions which was not previously supported by
`wasm-smith`.

Internally this refactors a number of pieces of code that work with
constant expressions to instead use a shared helper for generating
constant expressions. This takes into account subtyping and such to try
to generate interesting shapes of expressions when GC is enabled in
particular.