Replace custom "test harnesses" with `libtest-mimic` (#1450)
* Replace custom "test harnesses" with `libtest-mimic`

There's a number of tests suites in this repository which are based on
"drop a file here and it's a test now". Up to now these have been
powered by a custom "test harness" copied around which implements at
filtering mechanism but little else. The `libtest-mimic` crate provides
a `test`-crate-lookalike API which enables dynamically building the list
of tests at program start time (aka read the filesystem) and then
afterwards looks and feels similar to the `test` crate's CLI and UX.

This enables a nicer dev experience when running these tests and
additionally enables using runners like `cargo nextest` with this
repository.

* Don't be so fancy about running a single test

* Disable threads in CLI test