allow implementing world imports via adapter (#1459)
Previously, it was not possible to supply an adapter which implements
e.g. `wasi:cli/environment@0.2.0` such that `wit-component` would replace the
main (or library) module's import with the adapter's version.  This commit makes
that possible by prefering the adapter version over the import whenever there's
a match (i.e. an adapter with a name matching the WIT interface name and version).

This enables a form of lightweight virtualization in cases where we want to
replace imported functions with e.g. trapping stubs or a few lines of WAT --
without adding component composition to the mix.  This approach also works with
e.g. `wasmtime-py`, which as of this writing does not support composed
components.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>