Implement a `json-from-wast` subcommand (#1460)
* Implement a `wast2json` subcommand

This commit is an implementation of #1395 which brings wabt's
`wast2json` functionality to `wasm-tools`. This is implemented entirely
in the CLI since if using the crate it's already got all the
functionality necessary for this. Additionally this PR attempts to be
pretty close to `wast2json` from wabt in terms of structure and syntax.
The main difference is that wabt fills out the expected return types of
an operation when the operation isn't supposed to return (e.g.
`assert_trap`), but `wasm-tools` won't do that as we don't have that
information so easily available.

This is tested by adding `*.json` files to the `snapshots` directory
with example outputs from all known `*.wast` tests. Everything should be
included except for the `component-model` tests at this time which don't
have support in printing the args/results.

Closes #1395

* Fix tests on wasm

* Fix building just wast2json

* Generate the same test output on Windows

* Rename to `json-from-wast`

* Fix typo